+++ frontend/src/apis/deletes.ts
@@ -1,9 +1,7 @@
 import ApiClient from './apiClient';
-import { checkStatus } from './apiconfig';
 
 export const deleteCancelChamyo = async (moimId: number) => {
-  const response = await ApiClient.deleteWithAuth(`chamyo`, {
+await ApiClient.deleteWithAuth(`chamyo`, {
     moimId,
   });
-  await checkStatus(response);
 };
