+++ frontend/src/apis/gets.ts
@@ -10,12 +10,10 @@ import {
 } from './responseTypes';
 
 import ApiClient from './apiClient';
-import { checkStatus } from './apiconfig';
 import { Filter } from '@_components/MyMoimListFilters/MyMoimListFilters';
 
 export const getMoims = async (): Promise<MoimInfo[]> => {
   const response = await ApiClient.getWithAuth('moim');
-  checkStatus(response);
 
   const json: GetMoims = await response.json();
   return json.data.moims;
@@ -25,7 +23,6 @@ export const getMyFilteredMoims = async (
   filter: Filter['api'],
 ): Promise<MoimInfo[]> => {
   const response = await ApiClient.getWithAuth(`moim/mine?filter=${filter}`);
-  checkStatus(response);
 
   const json: GetMoims = await response.json();
   return json.data.moims;
@@ -33,7 +30,6 @@ export const getMyFilteredMoims = async (
 
 export const getMyZzimMoims = async (): Promise<MoimInfo[]> => {
   const response = await ApiClient.getWithAuth('moim/zzim');
-  checkStatus(response);
 
   const json: GetMoims = await response.json();
   return json.data.moims;
@@ -41,7 +37,6 @@ export const getMyZzimMoims = async (): Promise<MoimInfo[]> => {
 
 export const getMoim = async (moimId: number): Promise<MoimInfo> => {
   const response = await ApiClient.getWithAuth(`moim/${moimId}`);
-  checkStatus(response);
 
   const json: GetMoim = await response.json();
   return json.data;
@@ -54,7 +49,6 @@ export const getChat = async (
   const response = await ApiClient.getWithAuth(
     `chat?moimId=${moimId}&recentChatId=${recentChatId || 0}`,
   );
-  checkStatus(response);
 
   const json: GetChat = await response.json();
   return json.data.chats;
@@ -62,7 +56,6 @@ export const getChat = async (
 
 export const getMyMoims = async (): Promise<MoimInfo[]> => {
   const response = await ApiClient.getWithAuth(`moim/mine`);
-  checkStatus(response);
 
   const json: GetMoims = await response.json();
   return json.data.moims;
@@ -72,7 +65,6 @@ export const getChamyoMine = async (
   moimId: number,
 ): Promise<'MOIMER' | 'MOIMEE' | 'NON_MOIMEE'> => {
   const response = await ApiClient.getWithAuth(`chamyo/mine?moimId=${moimId}`);
-  checkStatus(response);
 
   const json: GetChamyoMine = await response.json();
   return json.data.role;
@@ -80,7 +72,6 @@ export const getChamyoMine = async (
 
 export const getZzimMine = async (moimId: number): Promise<boolean> => {
   const response = await ApiClient.getWithAuth(`zzim/mine?moimId=${moimId}`);
-  checkStatus(response);
 
   const json: GetZzimMine = await response.json();
   return json.data.isZzimed;
@@ -90,7 +81,6 @@ export const getChamyoAll = async (
   moimId: number,
 ): Promise<Participation[]> => {
   const response = await ApiClient.getWithAuth(`chamyo/all?moimId=${moimId}`);
-  checkStatus(response);
 
   const json: GetChamyoAll = await response.json();
   return json.data.chamyos;
